@@ -1,8 +1,9 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-
+    <TargetFramework>net7.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    
     <IsPackable>false</IsPackable>
   </PropertyGroup>
 
