@@ -1,10 +1,9 @@
 ﻿using System;
 
-namespace DotNetExtensions
+namespace DotNetExtensions;
+
+internal sealed class Void
 {
-    internal sealed class Void
-    {
-        private Void() { if (Instance != null) throw new InvalidOperationException(); }
-        public static readonly Void Instance = new Void();
-    }
+    private Void() { if (Instance != null) throw new InvalidOperationException(); }
+    public static readonly Void Instance = new();
 }
