@@ -1,9 +1,8 @@
 using Monads.TestAbstractions;
 
-namespace Monads.DataOps.Tests
+namespace Monads.DataOps.Tests;
+
+public partial class ClearableValueTests : TestsBase
 {
-    public partial class ClearableValueTests : TestsBase
-    {
-        public ClearableValueTests(TestFixture testFixture) : base(testFixture) { }
-    }
+    public ClearableValueTests(TestFixture testFixture) : base(testFixture) { }
 }
