@@ -1,19 +1,18 @@
 ﻿using System;
 
-namespace Monads.DataOps
+namespace Monads.DataOps;
+
+public static class Functions
 {
-    public static class Functions
+    public static T Id<T>(T arg) =>
+        arg;
+
+    public static T? NullableId<T>(T arg) where T : struct =>
+        arg;
+
+    public static T SideEffect<T>(T arg, Action<T> sideEffect)
     {
-        public static T Id<T>(T arg) =>
-            arg;
-
-        public static T? NullableId<T>(T arg) where T : struct =>
-            arg;
-
-        public static T SideEffect<T>(T arg, Action<T> sideEffect)
-        {
-            sideEffect(arg);
-            return arg;
-        }
+        sideEffect(arg);
+        return arg;
     }
 }
