@@ -1,8 +1,7 @@
-﻿namespace Monads.FluentAssertions
+﻿namespace Monads.FluentAssertions;
+
+public static class EitherAssertionExtensions
 {
-    public static class EitherAssertionExtensions
-    {
-        public static EitherAssertions<TLeft, TRight> Should<TLeft, TRight>(this Either<TLeft, TRight> actualValue) => new EitherAssertions<TLeft, TRight>(actualValue);
-        public static NullableEitherAssertions<TLeft, TRight> Should<TLeft, TRight>(this Either<TLeft, TRight>? actualValue) => new NullableEitherAssertions<TLeft, TRight>(actualValue);
-    }
+    public static EitherAssertions<TLeft, TRight> Should<TLeft, TRight>(this Either<TLeft, TRight> actualValue) => new(actualValue);
+    public static NullableEitherAssertions<TLeft, TRight> Should<TLeft, TRight>(this Either<TLeft, TRight>? actualValue) => new(actualValue);
 }
