@@ -1,8 +1,7 @@
-﻿namespace Monads.FluentAssertions
+﻿namespace Monads.FluentAssertions;
+
+public static class MaybeAssertionExtensions
 {
-    public static class MaybeAssertionExtensions
-    {
-        public static MaybeAssertions<T> Should<T>(this Maybe<T> actualValue) => new MaybeAssertions<T>(actualValue);
-        public static NullableMaybeAssertions<T> Should<T>(this Maybe<T>? actualValue) => new NullableMaybeAssertions<T>(actualValue);
-    }
+    public static MaybeAssertions<T> Should<T>(this Maybe<T> actualValue) => new(actualValue);
+    public static NullableMaybeAssertions<T> Should<T>(this Maybe<T>? actualValue) => new(actualValue);
 }
