@@ -1,9 +1,8 @@
 using Monads.TestAbstractions;
 
-namespace Monads.Tests
+namespace Monads.Tests;
+
+public partial class ResultExtensionsTests : TestsBase
 {
-    public partial class ResultExtensionsTests : TestsBase
-    {
-        public ResultExtensionsTests(TestFixture testFixture) : base(testFixture) { }
-    }
+    public ResultExtensionsTests(TestFixture testFixture) : base(testFixture) { }
 }
