@@ -1,9 +1,8 @@
 using Monads.TestAbstractions;
 
-namespace Monads.Tests
+namespace Monads.Tests;
+
+public partial class ResultTests : TestsBase
 {
-    public partial class ResultTests : TestsBase
-    {
-        public ResultTests(TestFixture testFixture) : base(testFixture) { }
-    }
+    public ResultTests(TestFixture testFixture) : base(testFixture) { }
 }
