@@ -1,12 +1,11 @@
-﻿namespace Monads
-{
-    public struct ResultOk<T>
-    {
-        public ResultOk(T result)
-        {
-            Result = result;
-        }
+﻿namespace Monads;
 
-        public T Result { get; }
+public struct ResultOk<T>
+{
+    public ResultOk(T result)
+    {
+        Result = result;
     }
+
+    public T Result { get; }
 }
